Debug view: Initial UI + Usage in MagicWeatherCompose (#1075)
### Description
Same as was added in iOS in
https://github.com/RevenueCat/purchases-ios/pull/2567. Includes the
changes in #1180

This PR adds a debug view in the Android SDK and uses it in the
`MagicWeatherCompose` sample app. This debug view is meant to be a
Jetpack compose function, even though later we might add convenience
accessors for people not using jetpack compose. It's extracted to a
different module so it can be published separately from the main SDK. It
also attempts to keep all the code in the debug source set, to try to
include as few code in the release variants.

The provided API is currently 2 composable functions:
- `DebugRevenueCatDebugScreen`: This is the actual UI that is part of
the debug screen. In case devs want to use it in a custom UI.
- `DebugRevenueCatBottomSheet`: This will display the debug screen as a
bottom sheet. This is what's used in the `MagicWeatherCompose` sample
app.

##### TODO
- [ ] Allow visualizing offerings in debug menu
- [ ] Support non-composable apps by wrapping it into a fragment, with
utility launch methods.
- [ ] Split library for debug and release
- [ ] Add UI tests
- [ ] Support publishing the new module as different libraries for debug
and release.
- [ ] Remove module local substitution once debugview library has been
published


https://github.com/RevenueCat/purchases-android/assets/808417/b3a953e4-b96f-4b06-8294-d2bcf9b66c7d