Optimize diagnostics file management (#1194)
### Description
This PR is the first part of a possible approach to avoid loading huge
files in memory in diagnostics at once. Basically, in here we change to
reading the file per lines instead of reading everything in one go.

#### Behavior changes
- Diagnostics will be Android 24+ only.
- An important change in behavior is that, instead of using the number
of events in the diagnostics file to set a limit, we use the file size.
So if on SDK configuration we detect that it's reached X size, we would
clear the whole file and start over.
- When we reach the file size limit, we remove the whole file, instead
of just the oldest events.
- We remove the properties from the `MAX_EVENTS_STORED_LIMIT_REACHED`
event since they don't make sense with the new approach. (This will
require backend changes)
- Now we will have a limit to the max number of diagnostics events in
memory and per request.