Improve offerings start log (#1225)
### Description
I noticed we had a very misleading log `Offerings updated from
network.`. It seemed to indicate that that offerings had finished
updating but it actually indicated that offerings started to be updated
from network. I changed it to `Start Offerings update from network.` and
also changed the location of the log so it happens in all places where
offerings are actually fetched from network.