Add `formattedPricePerMonth` helper to `StoreProduct` and `PricingPhase` (#1255)
### Description
While working on paywalls in #1251, we needed to calculate the price per
month for products to get the value of a variable.
This PR extracts that logic to the main SDK, providing an optional
`Locale` parameter which will be `Locale.getDefault()` by default. This
function has been added to the `PricingPhase` class and the
`StoreProduct` interface for convenience.

#### API Changes
- Added `fun formattedPricePerMonth(locale: Locale =
Locale.getDefault()): String?` to `StoreProduct`, `GoogleStoreProduct`
and `AmazonStoreProduct`.
- Added `fun formattedPricePerMonth(locale: Locale =
Locale.getDefault()): String` to `PricingPhase`.