Paywalls: Fix state update upon locale changes (#1287)
### Description
I noticed the locale changes weren't working as expected. It changed the
system driven locales but not our computed strings. This was because the
code that was meant to update the state upon locale changes was wrong.
We were using `remember` to store the last locales, but that doesn't
survive configuration changes like the view model does. So I moved the
logic to the view model.