Paywalls: Add support to launch paywall as activity (#1317)
### Description
Adds a `PaywallActivityLauncher` class that devs can instantiate in
their activity/fragment `onCreate` in order to launch the
`PaywallActivity` which is a wrapper around the full screen
`PaywallView`. Then, they receive the result in a `PaywallResultHandler`
interface.

Once they want to launch the activity, they can call the `launch` method
in that `PaywallActivityLauncher`.