Optimize billing client connection retries (#1300)
### Description
This is a possible approach to deal with #1288 
Seems like Google may be calling the `onBillingServiceDisconnected`
callback multiple times consecutively. In those cases we are scheduling
retries in the future for each of those calls which adds some overheard
to that bug in Google's side. In order to minimize that overhead, this
PR suggests having only a maximum of 1 retry scheduled, which should
reduce the overhead a lot.