`Paywalls`: `PaywallDialogOptions` no longer requires `dismissRequest` (#1386)
`PaywallDialog` works by presenting paywalls automatically based on a
provided condition. Because of this, it should also know how to dismiss
itself. I changed it to implement this automatically, but also provide a
way to optionally pass a block for it.

Also updated the docs to reflect the new APIs:
https://github.com/RevenueCat/revenuecat-docs/pull/450