Paywalls: Fix purchase button text flashing when changing selected package (#1416)
### Description
In some cases, the purchase button text doesn't change when selecting
different packages. However, we were still animating the text resulting
in a small flash in the purchase button when changing the selected
package.

This makes it so, if the purchase button text doesn't change between all
the packages, we don't animate it.