Paywalls: Add bundled font to paywall tester (#1415)
### Description
Using downloadable Google fonts when displaying the paywall has a small
delay where the default font is first used, then changed to the new font
the first time the font is used. While we could predownload the font
during the initial paywall tester app, this switches to using a bundled
font which is loaded immediately. This is the same font we were using
previously through google fonts.