Paywalls: Fix wrong view model data after configuration changes (#1417)
### Description
I noticed that after a configuration change (for example moving from
landscape to portrait), the close button didn't work as expected on the
dialogs. I tracked it down to the fact that the view model kept the
first version of the dismiss callback, so it wasn't calling the new one
created after the configuration change.

I decided to remove the current approach of using keys to create
individual view models and instead I'm adding a method to the view model
that gets called every time we try to get the view model. If we detect
that the PaywallOptions have changed, we will update it on the view
model, ensuring we have the latest version of the callbacks.