Fix RevenueCatUI API issues and add API tests  (#1433)
### Description
This adds API tests for the paywalls module. In the process of writing
those API tests I found 2 issues that I wanted to get fixed while the
API is experimental:
- The `requiredEntitlementIdentifier` parameter in the
`PaywallActivityLauncher.launchIfNeeded` method was after some optional
parameters, making it mandatory to use named parameters if you didn't
want to pass the other optional parameters.
- In Java, it wasn't possible to create a `PaywallFont` since it was
using value classes that are not available in Java.