Paywalls: Fix purchasing regression by providing real activity (#1467)
### Description
With #1418, we are overriding the context in order to provide our own
locale to make sure the screen has the same locale in the whole screen.
However, with the current implementation, the context becomes a
`ContextImpl` which can't be converted to the original activity. This
was causing a crash when starting the purchase process

This PR fixes the issue by providing a new `LocalActivity` local
composition that is performed before the overriden context takes place
so we can access the activity in our composable tree.