Use state listener in Amazon store (#1483)
### Description
We were not calling the BillingAbstract `stateListener` after it was
connected. This adds a call to make sure we call it and keep the same
behavior as on Google. Currently, we were just syncing pending purchases
in Google, which shouldn't matter much if we do it there for Amazon
since we're also doing on app foreground.