Initialize shared preferences lazily (#1489)
### Description
We detected some `DiskReadViolation` caused by initilizing shared
preferences. This should have a very low impact on initialization time,
but might be good to delay that so it's not caused until actually
needed. We currently have 3 shared preferences files, 2 of them are only
used later on and can be initialized later.