Avoid creating threads when dispatching tasks (#1496)
### Description
We were creating threads every time we added a task to the executor
service. This shouldn't be a problem if tasks ended early and not too
many tasks were in the queue, but it's not very efficient. This is
another implementation to propagate exceptions to the main thread so
they can be thrown and not swallowed when they happen on the thread
pool.