Refactor QueryPurchasesUseCase and adds QueryPurchasesByTypeUseCase (#1495)
We are using `BillingClient.queryPurchases` for `getPurchaseType`. I've
realized the easiest to reuse `QueryPurchasesUseCase` for both was to
split it into a `QueryPurchasesByTypeUseCase` that can get only
purchases of a certain type.

This has the consequence that in `QueryPurchasesUseCase` we are now
calling another `BillingClientUseCase` twice, meaning that we are
posting to the main thread twice. I actually think this is an advantage
so we split the work and treat the retries independently, but let me
know what you think.