Paywalls: Auto-close paywall activity if restore grants required entitlement identifier (#1507)
### Description
Users could potentially restore their purchases and get entitlements
granted but the paywall activity was not dismissing. This will add that
functionality in the case where users use the
`requiredEntitlementIdentifier` parameter to launch the activity.

This adds a new `PaywallResult` value, `Restored`. This will be returned
if the last action the user performed was a restore.