Paywalls: Move paywall view attributes parsing (#1514)
### Description
This is a possible approach to [this
comment](https://github.com/RevenueCat/purchases-android/pull/1510#pullrequestreview-1762883695).
Basically, there is a bunch of duplicated logic between `PaywallView`
and `PaywallFooterView`. This extracts part of it to a different class,
the logic related to parsing xml attributes.

We could try to create a superclass/interface that abstracts some of the
other methods... But that is relatively straightforward, and didn't want
to have to expose another type just for that.