Wait for tests to pass when deploying (#1521)
I noticed it's possible to deploy without waiting for all tests to pass.
It's probably an edge case but it's possible we merge something into
`main` that breaks unit tests then start a release from that broken
commit.

I added a new hold job to force the release in case we actually don't
want to wait for tests to pass. The name is pretty clear
`force-release-without-waiting-for-tests` so I think people will
understand that it's skipping tests.