PaywallActivityLauncher: Add callback indicating whether the paywall was displayed when presented conditionally to an entitlement identifier (#1542)
### Description
In the hybrid SDKs, we want to return a result from the
`presentPaywall`/`presentPaywallIfNeeded` methods with the result of the
paywall. However, we can't do that easily since `presentPaywallIfNeeded`
might not present the paywall at all, and there was no way to know
whether the paywall was displayed in that case until now.

This adds a new parameter to the `launchIfNeeded` methods that receive a
`requiredEntitlementIdentifier` that will indicate whether the paywall
was displayed. Note that this isn't added to the `launchIfNeeded` method
that received a block of code, since the client should know in that case
whether the paywall was displayed or not, depending on the result of
that callback.