Fix NullPointerException issues getting advertising id (#1550)
### Description
This was reported in #1548. I wasn't able to reproduce but from looking
around, seems that an update to the the latest version of the library
should fix it. However, the latest version is pretty old (from 1/2022).
Just in case, I added a catch to that block of code with a log, so it
doesn't crash.