Fix paywall dialog composable sizing issues (#1556)
### Description
As I was testing #1551, I noticed that the dialog size wasn't correct.
The width of the device sometimes was `MEDIUM` on a phone, causing it to
present the dialog in a narrowed version. After more testing, I noticed
that it was always one of the dimensions that were not `COMPACT` on my
emulator, so I decided to check both. If one of the dimensions is
`COMPACT`, we would present the dialog as full screen.

| Before  | After |
| ------------- | ------------- |
|
![image](https://github.com/RevenueCat/purchases-android/assets/808417/de5900ca-f0cc-48d1-8e32-64ea58819a7c)
|
![image](https://github.com/RevenueCat/purchases-android/assets/808417/c3a987a8-37b2-444c-81cc-4e8a16a98b44)
|