Add name property to StoreProduct (#1567)
### Description
Currently `StoreProduct`, `GoogleStoreProduct`, `AmazonStoreProduct` and
`TestStoreProduct` expose the `title` of the product. With Google, they
have 2 fields for products:
- The
[name](https://developer.android.com/reference/com/android/billingclient/api/ProductDetails#getName())
which actually has the product name
- The
[title](https://developer.android.com/reference/com/android/billingclient/api/ProductDetails#getTitle())
which has the product name + app name

We were only exposing the `title`. We probably should also expose the
`name`. This means that for `AmazonStoreProduct`, the title will be the
same as the `name`, but I think that's fine as long as it's documented.

Additionally, this changes our paywalls to use the `name` property
instead of the `title`, since it's more appropriate for the paywalls.