`Paywalls`: log error when creating `PaywallState.Error`
I'm currently debugging an issue with a developer, and realized that we have no way of knowing if the paywall goes into this state other than the error dialog being displayed.
This logs an error as well to make it easier to debug.

Example:
```
RevenueCatUI            com.revenuecat.paywall_tester        E  Paywall transitioned to error state: Error 10: Error performing request.
```