Fix `DEFERRED` purchases when passing old product id in invalid format (#1593)
`DEFERRED` is not currently supported in the latest SDK release, but on
6.9.5, the latest version it's supported on, the `DEFERRED` purchase
might not work if the developer gives as an old product id in the format
`productId:basePlanId`. Reason for it is that for deferred purchases, we
used the given old product id as the key of the callback, which doesn't
match the one returned once the purchase is completed, so the purchase
is successful, but the value is never returned.

This applies a patch so we make sure to use only the `productId` as the
`oldProductIdentifier` even if the developer gives us the other format
(which can be obtained from several places in our SDK). This will be
release as a hotfix and we will make hotfixes of the hybrids as well