Refactor `presentedOfferingIdentifier` into `presentedOfferingContext` object (#1612)
### Description
With some new incoming features, we will need to pass more context
information about where the package/product was obtained from. This PR
prepares for that by adding a new class `PresentedOfferingContext` which
right now only holds the `presentedOfferingIdentifier` but that we can
add more info later on that allows to pass more information to our
servers.