Paywalls: Improve reliability in case the state changes from loaded to a different state (#1620)
### Description
While this should not currently happen, in case the state changes from
`Loaded` to a different state, the current cast could potentially crash,
since the state is rendered during the out animation. This fixes that so
it doesn't crash and we just log an error instead.

This was reported in #1611 as potentially dangerous.