Add `getCurrentOffering(forPlacement: String)` to `Offerings` (#1604)
### Motivation

New function to get `Offering` by a placement identifier

### Description

New `getCurrentOfferingForPlacement(placementId: String): Offering?`
method on `Offerings`
- Added demo button for this in paywalls example app
- Added demo button for this in purchase tester app

```kotlin
// Example usage
offerings.getCurrentOfferingForPlacement("onboarding")?.let {
  showPaywall(it)
}
```

#### Logic flow

1. Returns an `Offering` if the placement identifier has a value
2. Checks if the placement identifier key exists in the `placements`
dictionary
    - If the key exists (and value is `null`)
1. Returns `null` if the placement identifier key exists in the
dictionary but has a `null` value (this is intentional by the developer
to NOT have an `Offering` for this placement identifier)
    - If the key doesn't exist
1. Return the `Offering` for `fallback_offering_id` (if its not null)

#### New placements object in offerings response

```js
{
    // other offerings stuff
    "placements": {
        "fallback_offering_id": "offering_a", // can be null
        "offering_ids_by_placement": {
            "onboarding_start": null,
            "onboarding_end": "offering_b",
        }
  }
}
```

#### Mapping `JSONObject.NULL` to `null`

Placements is the first thing that I think is allowed to have a null
value where JSONObject.toMap() was needed. This function, even though
allowed for an optional T type with Map<String, T>, it didn't' actually
properly map a value to null. Instead, it kept it as a JSONObject.NULL

Due to Kotlin typecasting, it wasn't really possible check if T was T or
T? so JSONObject.NULL couldn't be mapped safely to null.

This feels not as nice because its a whole new method, but its very
explicit in what it does... but I'm also happy if we have a better name
than this 😛

### Demo

#### Purchase Tester


https://github.com/RevenueCat/purchases-android/assets/401294/6384d93d-ef1c-44de-b151-299f612b7514

#### Paywall Tester


https://github.com/RevenueCat/purchases-android/assets/401294/3477a014-7303-42c2-aaf8-8211584e5ee7