Revert to previous androidx core version (1.8.0) (#1640)
### Description
In #1622, we updated to androidx core 1.12.0 which increased the target
version to 34. This however, brought some problems in some of the
hybrids: https://github.com/RevenueCat/purchases-capacitor/issues/211.

This update was to access some optional method that we can easily
implement ourselves, so we are reverting that update and implementing
the fallback for registering the receiver ourselves.

We will add additional checks to avoid inadvertently upgrading the
target version.