Make clearing diagnostics file happen during tracking (#1685)
### Description
We were clearing the diagnostics file during SDK configuration if we
detected the file to be too big. We could make this clearing happen
during tracking of events. Advantages of this:
- We make sure the file never gets too big 
- We don't miss extra events if we need to clear the file

This shouldn't be a problem performance-wise since we are doing all this
in a background thread + checking the file size is pretty quick.