Fix purchase tester offerings not reloading (#1715)
### Description
I've run into some issues with purchase tester for a while where the
offerings were not updated after changing screens or purchasing. This
simplifies the reloading so it loads always. There might be more
performant options, but considering this is a test app, I didn't want to
dedicate too much time to it.