Change android integration test to cover more testable use case (#1760)
### Description
We had a test that verifies that a token gets recognized and
entitlements given if the user is offline, then they recover, but that
test wasn't working correctly, since the token we used didn't grant any
entitlements (was an expired token).

After changing to a proper token, this test started failing because the
token was already assigned to a different user and we were posting it
with initiation source `unsynced_active_purchases`, which doesn't
transfer active purchases, so the entitlement was never granted. In
order to properly test that case, we would need a new token every time
we run the tests which we currently can't have, since testing with
Google Play is bad :(.

For now, I've changed the test to cover recovery after a restore, which
does work as expected. Happy to provide more details as needed!