Use trimmed API key (#1768)
### Description
I noticed that trusted entitlements may fail if passing an API key with
extra whitespaces. This is because it's part of the data we are
verifying and the backend uses the trimmed version so there is a
mismatch. This trims the API key we use in the SDK, so it matches what
the backend uses in those cases. Any real API key shouldn't contain any
whitespaces so I think it should be fine but lmk what you think