Update `compose-bom` to latest version (#1817)
### Description
We are getting some issues with Template7 of paywalls in RN and flutter
(it works fine on native):
-
https://community.revenuecat.com/sdks-51/revenuecat-paywall-component-causing-crash-on-android-nosuchmethoderror-in-animatedcontent-method-4956
- https://github.com/RevenueCat/purchases-flutter/issues/1145

From looking around at the issue, some people mentioned a library
incompatibility, so I updated the compose versions we're using and that
seems to solve the issue. After investigating, seems that the hybrids
are updating to a newer version of compose that is not compatible with
the version we were using in the native. We were also using an
experimental API (which is not experimental in the latest version)

I did a run through our UI and didn't notice any issues, but might be
good for someone else to double-check