Fix multiple paywalls display issues (#1854)
### Description
This is an approach to fix DENG-955

If trying to display multiple paywalls at the same time, all paywalls
would currently share the same view model. This caused that all of them
would be exactly the same, which was not great for some use cases. This
adds a key to the view model injection using the paywall options view
model hash code. This means we will have one view model for each
combination of paywall options parameters.


![image](https://github.com/user-attachments/assets/0d9d860e-9db9-4d51-a590-cf4de3fed42b)