Save SDK configuration on paywall activity so it can reconfigure automatically (#1872)
### Description
In here we store the latest known state for `Purchases` so we can
reconfigure when displaying the `PaywallActivity` and the SDK is not
configured. This may happen if the app is killed while displaying the
`PaywallActivity`, the developer is not configuring in the application
but on an activity and the user returns through the app through the
recent apps menu.

Additionally, it includes the changes from #1866 so we avoid
reconfiguring the SDK if the configuration parameters are the same.