Handle `IllegalArgumentException` when opening a Uri in the paywalls (#1904)
### Description
Looking at the
[docs](https://developer.android.com/reference/kotlin/androidx/compose/ui/platform/AndroidUriHandler#openUri(kotlin.String)),
the `openUri` method, may throw an `IllegalArgumentException` if the
link is invalid.

I haven't been able to reproduce this error, getting an
`ActivityNotFoundException` instead when the link is invalid... Seems
this may be something device related. In any case, we should also handle
these exceptions according to the docs and fail gracefully.