Add extra error cases for web purchases redemptions (#1899)
### Description
This adds some new result types to a web purchase redemption:
- InvalidToken
- AlreadyRedeemed
- Expired. This one includes additional information like the obfuscated
email where a new token was/will be sent and whether that email was
sent.

This falls behind the experimental APIs introduced in #1889 

Additionally, this renames `parseAsDeepLink` to
`parseAsWebPurchaseRedemption` and removes the `DeepLink` type. This is
a breaking change, but should be ok since the API is marked as
experimental and not usable still by any customers.