Increase integration test timeout (#1946)
### Description
Integration tests have been failing pretty frequently. From some tests,
it seems that verifying the purchase token in the backend after a while
takes a bit longer, which makes it frequently go over the timeout limit
of 5s set for purchasing. This increases the timeout so it has time to
finish the purchase on those situations.