[Paywalls V2] Add `Badge`'s `overlay` style layout (#2009)
### Description
This adds the layout for `Badge` to display correctly with the `overlay`
layout.

This PR moves the existing Stack layout inside another composable:
`MainStackComponent`. And at the top-level, we now decide what layout to
use depending on the badge style. In this PR, we are implementing the
first one: `overlay`.
<img width="595" alt="image"
src="https://github.com/user-attachments/assets/67150aac-c052-4ad6-a517-3071366f8392"
/>

---------

Co-authored-by: JayShortway <29483617+JayShortway@users.noreply.github.com>