[Trusted Entitlements] Do not clear CustomerInfo upon enabling Trusted Entitlements (#2049)
### Description
Until now, when devs enabled Trusted Entitlements, we would clear the
cached `CustomerInfo` and customers would need to refetch the latest
value from the server, which could potentially cause delays and/or
losing access temporarily.

With this PR, we won't be clearing the cached CustomerInfo anymore.
Instead, we would return the cached customer info, which will have a
`VerificationResult` of `NOT_REQUESTED`.

- [ ] Update docs