[Paywalls V2] Use original paywall fallback when trying to use Footer modes in a Components paywall (#2106)
### Description
Until now, if trying to present a components paywall as a footer, it
would display the paywall mostly full screen with some indeterminate
behavior.

After this, if the developer tries to present a paywall v2 as a
footer/condensed footer, we will fallback to the original paywall
fallback paywall, which uses template 2 right now.

Android equivalent of
https://github.com/RevenueCat/purchases-ios/pull/4667

---------

Co-authored-by: JayShortway <29483617+JayShortway@users.noreply.github.com>