[Paywalls V2] New overrides structure (#2120)
## Motivation

Migrating `overrides` from a structured object to an array of override
objects

## Description

### Previously

Overrides were a very structure object that had all of the different
types of states/conditions that allowed base properties of a component
to be overridden. However, this was not flexible when it came to
multiple being activated as it was missing for cases when multiples
types of conditions were being evaluated.

### Now

Overrides is an array that has an array of `conditions` and
`properties`. It will get applied in order from first to last IF all of
the `conditions` are true.

This allows the FE/BE to determine the orders of overrides from lowest
to higher priority so all of the SDKs/renderers don't need to duplicate
that logic.

Android equivalent of
https://github.com/RevenueCat/purchases-ios/pull/4705