Fixes `X` in promo offer not dismissing the promo (#2110)
Touching the `X` navigation button in the promotional offer closes the
whole Customer Center, this PR modifies that behavior and adds a way to
override what the X does. In the case of the promotional offer, it
should call `dismissPromotionalOffer` which effectively calls the main
action that the user triggered (like Cancel Purchase)

Also fixes another bug related to promotional offers, when the promo
offer can't be loaded, it wasn't triggering the main path either, it was
not doing anything.