[Paywalls V2] Fix stack clipping issues (#2126)
### Description
A number of components clip their contents so they don't go out of
bounds of the component itself. This is a problem for the
`StackComponent` since sometimes we want shadows of children or other
elements to extend beyond the bounds of the stack itself. An example of
this, is with overlaid badges.

This PR removes the clipping of the stack component itself except when
there is a Leading/Trailing EdgeToEdge badge or Nested badge, in which
we actually need to clip.

| Before | After |
| ---- | ---- |
|
![image](https://github.com/user-attachments/assets/dbc1aa52-4c18-47de-bb67-6366531d4aae)
|
![image](https://github.com/user-attachments/assets/c128a9a3-4480-4846-8033-f083fcd352ca)
|
| <img width="303" alt="image"
src="https://github.com/user-attachments/assets/87d952fc-4857-4d4a-857a-b78bdda5640f"
/> | <img width="311" alt="image"
src="https://github.com/user-attachments/assets/3934a977-a628-4e93-b9d3-859365480cd5"
/> |