[Paywalls V2] Make root stack component use all the available space by default (#2136)
### Description
Currently the backend sends the root stack component as having a width =
Fill, height = Fit, which is not fitting correctly with the way the main
layout in Android looks like. We believe the root component should take
all available space, so for now, this introduces a quick hack that makes
sure the paywall uses that space. This might change in the future
depending on how we want to support root stack sizes, but this looks how
it should in android with this change.

---------

Co-authored-by: JayShortway <29483617+JayShortway@users.noreply.github.com>