[Paywalls V2] Refactor to use a single ImageLoader singleton in RevenueCatUI (#2161)
### Description
This is an attempt to fix an issue where some icon images where failing
to load. Our theory is that the ImageLoader was getting deallocated,
since it seemed the connection was getting closed. I haven't been able
to reproduce though, so not sure if this will fix it. A couple things to
consider:
- We are only using a single ImageLoader when fetching the images, we
are still using multiple image loaders to predownload the images. Plan
to address this later
- I noticed for images when cache is disabled, we're not setting the
loader which in effect, would fallback to the "default" image loader. We
might want to fix that at some point as well.