Enable paywall tester minification (#2177)
### Description
We had minification disabled in paywall tester, which allowed us to miss
the issue fixed in #2176. This enables it so CI picks up future issues
with our UI module, which aren't caught by purchase tester. It also adds
a job to be run as part of the tests building the paywall tester app.