[Paywalls v2] Ignores already-consumed window insets. (#2326)
## Bug
The paywall was still applying window insets that were already consumed,
e.g. by a `Scaffold` or `AppBar`.

## Cause
I was under the impression that the `WindowInsets.systemBars` (and
equivalent) properties would only return not-yet-consumed insets, as
these are Composable properties. This turned out not to be the case.

## Fix
The fix is to use the `windowInsetsPadding` modifier instead. This
modifier will internally keep track of the consumed insets, and will
avoid applying them again.