Introduce purchases-bom package (#2339)
### Motivation

Introduce the `purchases-bom` package to ensure consistent dependency
management across multiple modules.

I noticed we were using version
[0.28.0](https://github.com/vanniktech/gradle-maven-publish-plugin/releases/tag/0.28.0)
of the Vanniktech plugin, which is built against Kotlin 1.9.20. This
caused compatibility issues with our current Kotlin Gradle Plugin
version 1.8.22, particularly when configuring the kotlin("jvm") plugin.
To resolve this, I downgraded the plugin to
[0.27.0](https://github.com/vanniktech/gradle-maven-publish-plugin/releases/tag/0.27.0),
which is compatible with Kotlin 1.8.22 that we're currently using.