[Paywalls V2] Renders template previews in a fixed resolution using Paparazzi (#2214)
## Description
This PR is the next step towards cross-platform paywall rendering
validation. It adds a `PaywallComponentsTemplatePreviewRecorder` which
renders our existing `PaywallComponentsTemplate_Preview` using
Paparazzi. This is done for 2 reasons:
- It allows us to pick a resolution and density, to be able to match the
other platforms.
- It gives us access to the screenshot files themselves, for further
processing.

Follow-up PRs will collect the screenshot files and push them to a
separate repo.

**Bonus:** this gives us a bit more coverage in the Android project too,
as these fixed-resolution screenshots are also uploaded to
[Emerge](https://www.emergetools.com/snapshot/compare/aa166da4-6534-434b-8215-5c033d08622f?tab=added).