[Paywalls V2] Improves fuzzy matching locale when the region doesn't match (#2355)
## Bug
We were still not always properly localizing the paywall when the region
doesn't match between the device and the paywall configuration.

## Cause
This is caused by the fact that in this scenario we use a LocaleId
containing only the language and script, without the region. If we then
try to find this LocaleId in the paywall's localizations map, we can't
find any, because the localizations map contains a region.

## Fix
We use the matched locale ID as it exists in the paywall config, instead
of a permutation.