[Paywalls v2] Records and pushes template screenshots to paywall-rendering-validation (#2365)
## Description
Adds a CI job and lane that will record paywall template screenshots and
push them to paywall-rendering-validation on every push to main. It then
opens a pull request. The folder structure in that repo for now is
`screenshots/templates/<offering-id>/<platform>.png`, but this can
easily be changed later.

Here's an example PR:
https://github.com/RevenueCat/paywall-rendering-validation/pull/3.